refactor: friendSummary 모델 세부 패키지 구성
