+++ Near/app/src/main/java/com/alarmy/near/data/repository/DefaultFriendRepository.kt
@@ -4,7 +4,7 @@ import com.alarmy.near.data.mapper.toModel
 import com.alarmy.near.data.mapper.toRequest
 import com.alarmy.near.model.Friend
 import com.alarmy.near.model.FriendRecord
-import com.alarmy.near.model.FriendSummary
+import com.alarmy.near.model.friendsummary.FriendSummary
 import com.alarmy.near.model.monthly.MonthlyFriend
 import com.alarmy.near.network.service.FriendService
 import kotlinx.coroutines.flow.Flow
