+++ Near/app/src/main/java/com/alarmy/near/model/friendsummary/ContactFrequencyLevel.kt
@@ -1,4 +1,4 @@
-package com.alarmy.near.model
+package com.alarmy.near.model.friendsummary
 
 enum class ContactFrequencyLevel {
     LOW,
