+++ Near/app/src/main/java/com/alarmy/near/model/friendsummary/FriendSummary.kt
@@ -1,4 +1,4 @@
-package com.alarmy.near.model
+package com.alarmy.near.model.friendsummary
 
 import androidx.compose.runtime.Immutable
